+++ src/main/java/it/gov/pagopa/fdr/controller/InternalOrganizationsController.java
@@ -1,6 +1,7 @@
 package it.gov.pagopa.fdr.controller;
 
-import it.gov.pagopa.fdr.controller.interfaces.IInternalOrganizationsController;
+import it.gov.pagopa.fdr.controller.interfaces.annotation.Re;
+import it.gov.pagopa.fdr.controller.interfaces.controller.IInternalOrganizationsController;
 import it.gov.pagopa.fdr.controller.model.flow.response.PaginatedFlowsResponse;
 import it.gov.pagopa.fdr.controller.model.flow.response.SingleFlowResponse;
 import it.gov.pagopa.fdr.controller.model.payment.response.PaginatedPaymentsResponse;
@@ -8,7 +9,6 @@ import it.gov.pagopa.fdr.service.FlowService;
 import it.gov.pagopa.fdr.service.PaymentService;
 import it.gov.pagopa.fdr.service.model.arguments.FindFlowsByFiltersArgs;
 import it.gov.pagopa.fdr.service.model.re.FdrActionEnum;
-import it.gov.pagopa.fdr.util.Re;
 import java.time.Instant;
 
 public class InternalOrganizationsController implements IInternalOrganizationsController {
