+++ src/main/java/it/gov/pagopa/fdr/controller/interfaces/controller/IInternalOrganizationsController.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.fdr.controller.interfaces;
+package it.gov.pagopa.fdr.controller.interfaces.controller;
 
 import it.gov.pagopa.fdr.controller.model.flow.response.PaginatedFlowsResponse;
 import it.gov.pagopa.fdr.controller.model.flow.response.SingleFlowResponse;
@@ -22,15 +22,15 @@ import org.eclipse.microprofile.openapi.annotations.responses.APIResponse;
 import org.eclipse.microprofile.openapi.annotations.responses.APIResponses;
 import org.eclipse.microprofile.openapi.annotations.tags.Tag;
 
-@Path(ControllerConstants.URL_CONTROLLER_ORGANIZATIONS)
+@Path(ControllerConstants.URL_CONTROLLER_INTERNAL_ORGANIZATIONS)
 @Consumes("application/json")
 @Produces("application/json")
-@Tag(name = "Organizations", description = "Organizations operations")
-public interface IOrganizationsController {
+@Tag(name = "Internal Organizations", description = "Organizations operations")
+public interface IInternalOrganizationsController {
 
   @GET
   @Operation(
-      operationId = "getAllPublishedFlows",
+      operationId = "getAllPublishedFlowsForInternalUse",
       summary = "Get all fdr published",
       description = "Get all fdr published")
   @APIResponses(
@@ -46,7 +46,7 @@ public interface IOrganizationsController {
                     mediaType = MediaType.APPLICATION_JSON,
                     schema = @Schema(implementation = PaginatedFlowsResponse.class)))
       })
-  PaginatedFlowsResponse getAllPublishedFlows(
+  PaginatedFlowsResponse getAllPublishedFlowsForInternalUse(
       @PathParam(ControllerConstants.PARAMETER_ORGANIZATION) @Pattern(regexp = "^(.{1,35})$")
           String organizationId,
       @QueryParam(ControllerConstants.PARAMETER_PSP) @Pattern(regexp = "^(.{1,35})$") String pspId,
@@ -62,7 +62,10 @@ public interface IOrganizationsController {
 
   @GET
   @Path(ControllerConstants.URL_API_GET_SINGLE_FLOW)
-  @Operation(operationId = "getSinglePublishedFlow", summary = "Get fdr", description = "Get fdr")
+  @Operation(
+      operationId = "getSingleFlowForInternalUse",
+      summary = "Get fdr",
+      description = "Get fdr by id but not payments")
   @APIResponses(
       value = {
         @APIResponse(ref = "#/components/responses/InternalServerError"),
@@ -76,16 +79,17 @@ public interface IOrganizationsController {
                     mediaType = MediaType.APPLICATION_JSON,
                     schema = @Schema(implementation = SingleFlowResponse.class)))
       })
-  SingleFlowResponse getSinglePublishedFlow(
-      @PathParam(ControllerConstants.PARAMETER_ORGANIZATION) String organizationId,
+  SingleFlowResponse getSingleFlowForInternalUse(
+      @PathParam(ControllerConstants.PARAMETER_ORGANIZATION) @Pattern(regexp = "^(.{1,35})$")
+          String organizationId,
       @PathParam(ControllerConstants.PARAMETER_FDR) String flowName,
       @PathParam(ControllerConstants.PARAMETER_REVISION) Long revision,
-      @PathParam(ControllerConstants.PARAMETER_PSP) String psp);
+      @PathParam(ControllerConstants.PARAMETER_PSP) String pspId);
 
   @GET
   @Path(ControllerConstants.URL_API_GET_FLOW_PAYMENTS)
   @Operation(
-      operationId = "getPaymentsFromPublishedFlow",
+      operationId = "getFlowPaymentsForInternalUse",
       summary = "Get payments of fdr",
       description = "Get payments of fdr")
   @APIResponses(
@@ -101,8 +105,9 @@ public interface IOrganizationsController {
                     mediaType = MediaType.APPLICATION_JSON,
                     schema = @Schema(implementation = PaginatedPaymentsResponse.class)))
       })
-  PaginatedPaymentsResponse getPaymentsFromPublishedFlow(
-      @PathParam(ControllerConstants.PARAMETER_ORGANIZATION) String organizationId,
+  PaginatedPaymentsResponse getFlowPaymentsForInternalUse(
+      @PathParam(ControllerConstants.PARAMETER_ORGANIZATION) @Pattern(regexp = "^(.{1,35})$")
+          String organizationId,
       @PathParam(ControllerConstants.PARAMETER_FDR) String flowName,
       @PathParam(ControllerConstants.PARAMETER_REVISION) Long revision,
       @PathParam(ControllerConstants.PARAMETER_PSP) String pspId,
