+++ src/main/java/it/gov/pagopa/fdr/controller/interfaces/controller/IInternalPspsController.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.fdr.controller.interfaces;
+package it.gov.pagopa.fdr.controller.interfaces.controller;
 
 import it.gov.pagopa.fdr.controller.model.common.response.GenericResponse;
 import it.gov.pagopa.fdr.controller.model.flow.request.CreateFlowRequest;
