+++ src/main/java/it/gov/pagopa/fdr/controller/interfaces/controller/IOrganizationsController.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.fdr.controller.interfaces;
+package it.gov.pagopa.fdr.controller.interfaces.controller;
 
 import it.gov.pagopa.fdr.controller.model.flow.response.PaginatedFlowsResponse;
 import it.gov.pagopa.fdr.controller.model.flow.response.SingleFlowResponse;
@@ -22,15 +22,15 @@ import org.eclipse.microprofile.openapi.annotations.responses.APIResponse;
 import org.eclipse.microprofile.openapi.annotations.responses.APIResponses;
 import org.eclipse.microprofile.openapi.annotations.tags.Tag;
 
-@Path(ControllerConstants.URL_CONTROLLER_INTERNAL_ORGANIZATIONS)
+@Path(ControllerConstants.URL_CONTROLLER_ORGANIZATIONS)
 @Consumes("application/json")
 @Produces("application/json")
-@Tag(name = "Internal Organizations", description = "Organizations operations")
-public interface IInternalOrganizationsController {
+@Tag(name = "Organizations", description = "Organizations operations")
+public interface IOrganizationsController {
 
   @GET
   @Operation(
-      operationId = "getAllPublishedFlowsForInternalUse",
+      operationId = "getAllPublishedFlows",
       summary = "Get all fdr published",
       description = "Get all fdr published")
   @APIResponses(
@@ -46,7 +46,7 @@ public interface IInternalOrganizationsController {
                     mediaType = MediaType.APPLICATION_JSON,
                     schema = @Schema(implementation = PaginatedFlowsResponse.class)))
       })
-  PaginatedFlowsResponse getAllPublishedFlowsForInternalUse(
+  PaginatedFlowsResponse getAllPublishedFlows(
       @PathParam(ControllerConstants.PARAMETER_ORGANIZATION) @Pattern(regexp = "^(.{1,35})$")
           String organizationId,
       @QueryParam(ControllerConstants.PARAMETER_PSP) @Pattern(regexp = "^(.{1,35})$") String pspId,
@@ -62,10 +62,7 @@ public interface IInternalOrganizationsController {
 
   @GET
   @Path(ControllerConstants.URL_API_GET_SINGLE_FLOW)
-  @Operation(
-      operationId = "getSingleFlowForInternalUse",
-      summary = "Get fdr",
-      description = "Get fdr by id but not payments")
+  @Operation(operationId = "getSinglePublishedFlow", summary = "Get fdr", description = "Get fdr")
   @APIResponses(
       value = {
         @APIResponse(ref = "#/components/responses/InternalServerError"),
@@ -79,17 +76,16 @@ public interface IInternalOrganizationsController {
                     mediaType = MediaType.APPLICATION_JSON,
                     schema = @Schema(implementation = SingleFlowResponse.class)))
       })
-  SingleFlowResponse getSingleFlowForInternalUse(
-      @PathParam(ControllerConstants.PARAMETER_ORGANIZATION) @Pattern(regexp = "^(.{1,35})$")
-          String organizationId,
+  SingleFlowResponse getSinglePublishedFlow(
+      @PathParam(ControllerConstants.PARAMETER_ORGANIZATION) String organizationId,
       @PathParam(ControllerConstants.PARAMETER_FDR) String flowName,
       @PathParam(ControllerConstants.PARAMETER_REVISION) Long revision,
-      @PathParam(ControllerConstants.PARAMETER_PSP) String pspId);
+      @PathParam(ControllerConstants.PARAMETER_PSP) String psp);
 
   @GET
   @Path(ControllerConstants.URL_API_GET_FLOW_PAYMENTS)
   @Operation(
-      operationId = "getFlowPaymentsForInternalUse",
+      operationId = "getPaymentsFromPublishedFlow",
       summary = "Get payments of fdr",
       description = "Get payments of fdr")
   @APIResponses(
@@ -105,9 +101,8 @@ public interface IInternalOrganizationsController {
                     mediaType = MediaType.APPLICATION_JSON,
                     schema = @Schema(implementation = PaginatedPaymentsResponse.class)))
       })
-  PaginatedPaymentsResponse getFlowPaymentsForInternalUse(
-      @PathParam(ControllerConstants.PARAMETER_ORGANIZATION) @Pattern(regexp = "^(.{1,35})$")
-          String organizationId,
+  PaginatedPaymentsResponse getPaymentsFromPublishedFlow(
+      @PathParam(ControllerConstants.PARAMETER_ORGANIZATION) String organizationId,
       @PathParam(ControllerConstants.PARAMETER_FDR) String flowName,
       @PathParam(ControllerConstants.PARAMETER_REVISION) Long revision,
       @PathParam(ControllerConstants.PARAMETER_PSP) String pspId,
